+++ spec/resolver_spec.rb
@@ -2,53 +2,6 @@ $:.unshift(File.dirname(__FILE__))
 
 require 'spec_helper'
 
-describe "RadioDNS::Service" do
-  before(:each) do
-    @service = RadioDNS::Service.new('rdns.musicradio.com')
-  end
-
-  it "has a cname" do
-    assert @service.respond_to? :cname
-  end
-
-  it "looks up radiovis application" do
-    @service.expects(:application).with(:radiovis).returns(nil)
-    @service.radiovis
-  end
-
-  it "looks up radiotag application" do
-    @service.expects(:application).with(:radiotag).returns(nil)
-    @service.radiotag
-  end
-
-  it "looks up radioepg application" do
-    @service.expects(:application).with(:radioepg).returns(nil)
-    @service.radioepg
-  end
-
-  it "has an array of supported applications" do
-    @service.expects(:application).with(:radioepg).returns(:epg)
-    @service.expects(:application).with(:radiovis).returns(:vis)
-    @service.expects(:application).with(:radiotag).returns(:tag)
-    assert_equal [:tag, :epg, :vis], @service.applications
-  end
-
-  it "performs SVR lookups" do
-    mock_resolver = mock()
-    mock_resource = mock()
-    mock_resolver.expects(:getresource).
-      with('_some_application._tcp.rdns.musicradio.com', Resolv::DNS::Resource::IN::SRV).once.
-      returns(mock_resource)
-    mock_resource.expects(:port).returns(1234)
-    mock_resource.expects(:target).returns("rdns.musicradio.com")
-    Resolv::DNS.expects(:new).returns(mock_resolver)
-
-    application = @service.application(:some_application)
-
-    assert_equal "rdns.musicradio.com", application.host
-    assert_equal 1234, application.port
-  end
-end
 
 describe "RadioDNS::Resolver" do
   describe "resolve" do
@@ -62,6 +15,7 @@ describe "RadioDNS::Resolver" do
 
       Resolv::DNS.expects(:new).returns(mock_resolver)
     end
+    
     it "should query radiodns.org" do
       service = RadioDNS::Resolver.resolve('09580.c586.ce1.fm.radiodns.org')
       assert_equal 'rdns.musicradio.com', service.cname
@@ -81,16 +35,6 @@ describe "RadioDNS::Resolver" do
 
   describe "construct_fqdn" do
     describe "for FM/VHF bearer" do
-      it "should construct a fqdn when freq is supplied as a Float" do
-        params = {
-          :bearer => 'fm',
-          :ecc => 'ce1',
-          :pi => 'c585',
-          :freq => 95.8
-        }
-        fqdn = RadioDNS::Resolver.construct_fqdn(params)
-        assert_equal '09580.c585.ce1.fm.radiodns.org', fqdn
-      end
       it "should construct a fqdn when ecc is supplied" do
         params = {
           :bearer => 'fm',
@@ -101,6 +45,7 @@ describe "RadioDNS::Resolver" do
         fqdn = RadioDNS::Resolver.construct_fqdn(params)
         assert_equal '09580.c585.ce1.fm.radiodns.org', fqdn
       end
+      
       it "should construct a fqdn when country is supplied" do
         params = {
           :bearer => 'fm',
@@ -111,6 +56,7 @@ describe "RadioDNS::Resolver" do
         fqdn = RadioDNS::Resolver.construct_fqdn(params)
         assert_equal '09580.c585.gb.fm.radiodns.org', fqdn
       end
+      
       it "should raise when country and ecc is supplied" do
         params = {
           :bearer => 'fm',
@@ -121,6 +67,7 @@ describe "RadioDNS::Resolver" do
         }
         assert_raises(ArgumentError) {RadioDNS::Resolver.construct_fqdn(params)}
       end
+      
       it "should raise when neither country nor ecc is supplied" do
         params = {
           :bearer => 'fm',
@@ -130,6 +77,7 @@ describe "RadioDNS::Resolver" do
         assert_raises(ArgumentError) {RadioDNS::Resolver.construct_fqdn(params)}
       end
     end
+    
     describe "for DAB bearer" do
       it "should construct a fqdn" do
         params = {
@@ -142,12 +90,14 @@ describe "RadioDNS::Resolver" do
         fqdn = RadioDNS::Resolver.construct_fqdn(params)
         assert_equal 'scids.sid.eid.ecc.dab.radiodns.org', fqdn
       end
+      
       it "should raise if parameters are missing" do
         params = {
           :bearer => 'dab',
         }
         assert_raises(ArgumentError) {RadioDNS::Resolver.construct_fqdn(params)}
       end
+      
       it "should prepend appty-uatype if provided" do
         params = {
           :bearer => 'dab',
@@ -160,6 +110,7 @@ describe "RadioDNS::Resolver" do
         fqdn = RadioDNS::Resolver.construct_fqdn(params)
         assert_equal 'appty-uatype.scids.sid.eid.ecc.dab.radiodns.org', fqdn
       end
+      
       it "should prepend pa if provided" do
         params = {
           :bearer => 'dab',
@@ -173,6 +124,7 @@ describe "RadioDNS::Resolver" do
         assert_equal 'pa.scids.sid.eid.ecc.dab.radiodns.org', fqdn
       end
     end
+    
     describe "for DRM bearer" do
       it "should construct a fqdn" do
         params = {
@@ -189,6 +141,7 @@ describe "RadioDNS::Resolver" do
         assert_raises(ArgumentError) {RadioDNS::Resolver.construct_fqdn(params)}
       end
     end
+    
     describe "for AMSS bearer" do
       it "should construct a fqdn" do
         params = {
@@ -205,6 +158,7 @@ describe "RadioDNS::Resolver" do
         assert_raises(ArgumentError) {RadioDNS::Resolver.construct_fqdn(params)}
       end
     end
+    
     describe "for hd bearer" do
       it "should construct a fqdn" do
         params = {
